+++ RELEASE_NOTES.md
@@ -3,8 +3,14 @@ Release Notes
 
 ## Next Release
 
+**New**
+
 - `Row.columns` returns the names of columns in the row.
 - `Row.databaseValues` returns the database values in the row.
+
+**Breaking changes**
+
+- `RowModel.edited` has been renamed `RowModel.databaseEdited`.
 
 
 ## 0.11.0